feat: upgrade to numen v0.3.0 and handle rate refresh
